Added filters

+++ Project/src/main/java/com/finance/financeapp/ejb/Gasto.java
@@ -17,7 +17,7 @@ public class Gasto implements Serializable {
 	private int idGasto;
 	
 	private String lugar;
-	private double monto;
+	private float monto;
 	private String descripcion;
 	private String fecha;
 	
@@ -56,14 +56,14 @@ public class Gasto implements Serializable {
 	/**
 	 * @return the monto
 	 */
-	public double getMonto() {
+	public float getMonto() {
 		return monto;
 	}
 
 	/**
 	 * @param monto the monto to set
 	 */
-	public void setMonto(double monto) {
+	public void setMonto(float monto) {
 		this.monto = monto;
 	}
